feat(storybook): color picker args for Skjermbilder-stories
Flytter delte token-helpers til packages/jokul/stories/colorTokens.tsx så
de kan importeres både fra .storybook og fra stories. Legger til
colorTokenArgTypes, emptyColorTokenArgs og colorArgsDecorator i alle
stories under "Skjermbilder", slik at tokens kan overstyres via Storybook
Controls per story.
